pyproject.toml
[tool.poetry]
name = "test_queue"
version = "0.1.0"
description = ""
authors = ["VIktor Vyunikov <vvyunnikov@bki-okb.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
aio-pika = "^7.1.2"
tortoise-orm = "^0.19.0"
asyncpg = "^0.25.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
